GODRIVER-2617 WIP
