+++ mongo/integration/mtest/setup.go
@@ -71,7 +71,15 @@ func setupClient(opts *options.ClientOptions) (*mongo.Client, error) {
 // Setup initializes the current testing context.
 // This function must only be called one time and must be called before any tests run.
 func Setup(setupOpts ...*SetupOptions) error {
-	opts := MergeSetupOptions(setupOpts...)
+	opts := NewSetupOptions()
+	for _, opt := range setupOpts {
+		if opt == nil {
+			continue
+		}
+		if opt.URI != nil {
+			opts.URI = opt.URI
+		}
+	}
 
 	var uri string
 	var err error
