+++ mongo/integration/mtest/setup_options.go
@@ -23,20 +23,3 @@ func (so *SetupOptions) SetURI(uri string) *SetupOptions {
 	so.URI = &uri
 	return so
 }
-
-// MergeSetupOptions combines the given *SetupOptions into a single *Options in a last one wins fashion.
-//
-// Deprecated: Merging options structs will not be supported in Go Driver 2.0. Users should create a
-// single options struct instead.
-func MergeSetupOptions(opts ...*SetupOptions) *SetupOptions {
-	op := NewSetupOptions()
-	for _, opt := range opts {
-		if opt == nil {
-			continue
-		}
-		if opt.URI != nil {
-			op.URI = opt.URI
-		}
-	}
-	return op
-}
