+++ mongo/integration/unified/options.go
@@ -25,21 +25,3 @@ func (op *Options) SetRunKillAllSessions(killAllSessions bool) *Options {
 	op.RunKillAllSessions = &killAllSessions
 	return op
 }
-
-// MergeOptions combines the given *Options into a single *Options in a last one wins fashion.
-//
-// Deprecated: Merging options structs will not be supported in Go Driver 2.0. Users should create a
-// single options struct instead.
-func MergeOptions(opts ...*Options) *Options {
-	op := NewOptions()
-	for _, opt := range opts {
-		if opt == nil {
-			continue
-		}
-		if opt.RunKillAllSessions != nil {
-			op.RunKillAllSessions = opt.RunKillAllSessions
-		}
-	}
-
-	return op
-}
