+++ mongo/options/aggregateoptions.go
@@ -136,49 +136,3 @@ func (ao *AggregateOptions) SetCustom(c bson.M) *AggregateOptions {
 	ao.Custom = c
 	return ao
 }
-
-// MergeAggregateOptions combines the given AggregateOptions instances into a single AggregateOptions in a last-one-wins
-// fashion.
-//
-// Deprecated: Merging options structs will not be supported in Go Driver 2.0. Users should create a
-// single options struct instead.
-func MergeAggregateOptions(opts ...*AggregateOptions) *AggregateOptions {
-	aggOpts := Aggregate()
-	for _, ao := range opts {
-		if ao == nil {
-			continue
-		}
-		if ao.AllowDiskUse != nil {
-			aggOpts.AllowDiskUse = ao.AllowDiskUse
-		}
-		if ao.BatchSize != nil {
-			aggOpts.BatchSize = ao.BatchSize
-		}
-		if ao.BypassDocumentValidation != nil {
-			aggOpts.BypassDocumentValidation = ao.BypassDocumentValidation
-		}
-		if ao.Collation != nil {
-			aggOpts.Collation = ao.Collation
-		}
-		if ao.MaxTime != nil {
-			aggOpts.MaxTime = ao.MaxTime
-		}
-		if ao.MaxAwaitTime != nil {
-			aggOpts.MaxAwaitTime = ao.MaxAwaitTime
-		}
-		if ao.Comment != nil {
-			aggOpts.Comment = ao.Comment
-		}
-		if ao.Hint != nil {
-			aggOpts.Hint = ao.Hint
-		}
-		if ao.Let != nil {
-			aggOpts.Let = ao.Let
-		}
-		if ao.Custom != nil {
-			aggOpts.Custom = ao.Custom
-		}
-	}
-
-	return aggOpts
-}
