+++ mongo/options/autoencryptionoptions.go
@@ -162,49 +162,3 @@ func (a *AutoEncryptionOptions) SetBypassQueryAnalysis(bypass bool) *AutoEncrypt
 	a.BypassQueryAnalysis = &bypass
 	return a
 }
-
-// MergeAutoEncryptionOptions combines the argued AutoEncryptionOptions in a last-one wins fashion.
-//
-// Deprecated: Merging options structs will not be supported in Go Driver 2.0. Users should create a
-// single options struct instead.
-func MergeAutoEncryptionOptions(opts ...*AutoEncryptionOptions) *AutoEncryptionOptions {
-	aeo := AutoEncryption()
-	for _, opt := range opts {
-		if opt == nil {
-			continue
-		}
-
-		if opt.KeyVaultClientOptions != nil {
-			aeo.KeyVaultClientOptions = opt.KeyVaultClientOptions
-		}
-		if opt.KeyVaultNamespace != "" {
-			aeo.KeyVaultNamespace = opt.KeyVaultNamespace
-		}
-		if opt.KmsProviders != nil {
-			aeo.KmsProviders = opt.KmsProviders
-		}
-		if opt.SchemaMap != nil {
-			aeo.SchemaMap = opt.SchemaMap
-		}
-		if opt.BypassAutoEncryption != nil {
-			aeo.BypassAutoEncryption = opt.BypassAutoEncryption
-		}
-		if opt.ExtraOptions != nil {
-			aeo.ExtraOptions = opt.ExtraOptions
-		}
-		if opt.TLSConfig != nil {
-			aeo.TLSConfig = opt.TLSConfig
-		}
-		if opt.EncryptedFieldsMap != nil {
-			aeo.EncryptedFieldsMap = opt.EncryptedFieldsMap
-		}
-		if opt.BypassQueryAnalysis != nil {
-			aeo.BypassQueryAnalysis = opt.BypassQueryAnalysis
-		}
-		if opt.HTTPClient != nil {
-			aeo.HTTPClient = opt.HTTPClient
-		}
-	}
-
-	return aeo
-}
