+++ mongo/options/changestreamoptions.go
@@ -153,55 +153,3 @@ func (cso *ChangeStreamOptions) SetCustomPipeline(cp bson.M) *ChangeStreamOption
 	cso.CustomPipeline = cp
 	return cso
 }
-
-// MergeChangeStreamOptions combines the given ChangeStreamOptions instances into a single ChangeStreamOptions in a
-// last-one-wins fashion.
-//
-// Deprecated: Merging options structs will not be supported in Go Driver 2.0. Users should create a
-// single options struct instead.
-func MergeChangeStreamOptions(opts ...*ChangeStreamOptions) *ChangeStreamOptions {
-	csOpts := ChangeStream()
-	for _, cso := range opts {
-		if cso == nil {
-			continue
-		}
-		if cso.BatchSize != nil {
-			csOpts.BatchSize = cso.BatchSize
-		}
-		if cso.Collation != nil {
-			csOpts.Collation = cso.Collation
-		}
-		if cso.Comment != nil {
-			csOpts.Comment = cso.Comment
-		}
-		if cso.FullDocument != nil {
-			csOpts.FullDocument = cso.FullDocument
-		}
-		if cso.FullDocumentBeforeChange != nil {
-			csOpts.FullDocumentBeforeChange = cso.FullDocumentBeforeChange
-		}
-		if cso.MaxAwaitTime != nil {
-			csOpts.MaxAwaitTime = cso.MaxAwaitTime
-		}
-		if cso.ResumeAfter != nil {
-			csOpts.ResumeAfter = cso.ResumeAfter
-		}
-		if cso.ShowExpandedEvents != nil {
-			csOpts.ShowExpandedEvents = cso.ShowExpandedEvents
-		}
-		if cso.StartAtOperationTime != nil {
-			csOpts.StartAtOperationTime = cso.StartAtOperationTime
-		}
-		if cso.StartAfter != nil {
-			csOpts.StartAfter = cso.StartAfter
-		}
-		if cso.Custom != nil {
-			csOpts.Custom = cso.Custom
-		}
-		if cso.CustomPipeline != nil {
-			csOpts.CustomPipeline = cso.CustomPipeline
-		}
-	}
-
-	return csOpts
-}
